+++ libs/aws/sabi-s3/src/actions/put_object/request/file_request.rs
@@ -4,7 +4,7 @@ use crate::core::request::{RequestResource, ResourceLoader};
 use crate::core::verbs::HasObjectKey;
 use sabi_core::auth::v4::canonical::HashedPayload;
 use sabi_core::auth::v4::chrono::now;
-use sabi_core::http::header::ContentType;
+use sabi_core::http::request::header::ContentType;
 use sabi_core::index::RegionCode;
 use sabi_core::io::stream::bytes_stream;
 use std::error::Error;
