+++ libs/aws/sabi-s3/src/core/request/request_resource.rs
@@ -2,7 +2,7 @@ use crate::core;
 use chrono::{DateTime, Utc};
 use reqwest::Body;
 use sabi_core::auth::v4::canonical::HashedPayload;
-use sabi_core::http::header::ContentType;
+use sabi_core::http::request::header::ContentType;
 use sabi_core::index::RegionCode;
 
 pub struct RequestResource<'a> {
