+++ libs/sabi-core/src/auth/v4/canonical/mod.rs
@@ -25,7 +25,7 @@ pub use signed_headers::SignedHeaders;
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::http::RichHeaderMap;
+    use crate::http::request::RichHeaderMap;
     use crate::SabiResult;
     use http::{HeaderMap, Method};
     use url::Url;
