+++ libs/sabi-core/src/http/mod.rs
@@ -1,5 +1,1 @@
-mod headers;
-pub use headers::HeaderFragment;
-pub use headers::RichHeaderMap;
-pub use headers::ToHeaderFragment;
-pub mod header;
+pub mod request;
